Se modificó el carousel de videos y se resolvió el bug
