--- src/components/ChargeVideos/styles.js
@@ -1,32 +1,0 @@
-import styled from "styled-components"
-
-export const StyledLabelChargeVideos = styled.label`
-  font-family: Roboto;
-  font-weight: 700;
-  font-size: 16px;
-  color: #FBFCF9;
-  margin: 10px;
-`
-export const StyledFormChargeVideos = styled.form`
-  position: relative;
-  margin-left:15%;
-  margin-bottom: 4%;
-`
-
-export const StyledInputVideos = styled.input`
-  margin: 3%;
-`
-
-export const StyledChargeVideoSection = styled.section`
-  width: 100%;
-  height: 50%;
-`
-
-export const StyledLabelVideoUploader = styled.label`
-  width: 100%;
-  font-family: Roboto;
-  font-weight: 700;
-  font-size: 16px;
-  color: #FBFCF9;
-  margin-left: 37%;
-`
