+++ src/components/CoachProfileForm/styles.css
@@ -1,7 +1,6 @@
 .right-arrow {
   position: absolute;
-  margin-left: 130%;
-  margin-bottom: 15%;
+  margin-left: 120%;
   font-size: 3rem;
   color: #000;
   z-index: 10;
@@ -11,8 +10,7 @@
 
 .left-arrow {
   position: absolute;
-  margin-right: 130%;
-  margin-bottom: 15%;
+  margin-right: 120%;
   font-size: 3rem;
   color: #000;
   z-index: 10;
