+++ src/components/CoachProfileForm/styles.js
@@ -112,3 +112,47 @@ export const StyledPicture = styled.div`
   display: ${props => props.picture ? "grid" : "flex"};
   justify-content: center;
 `
+
+//Estilos de la seccion de carga de videos
+export const StyledLabelChargeVideos = styled.label`
+  font-family: Roboto;
+  font-weight: 700;
+  font-size: 16px;
+  color: #FBFCF9;
+  margin: 10px;
+`
+export const StyledFormChargeVideos = styled.form`
+  position: relative;
+  margin-left:25%;
+  margin-bottom: 4%;
+`
+
+export const StyledInputVideos = styled.input`
+  margin: 3%;
+`
+
+export const StyledChargeVideoSection = styled.section`
+  width: 100%;
+  height: 50%;
+`
+
+//estilos de la sección de carousel videos
+export const SliderContentItem = styled.section`
+  position: relative;
+  height: 100hv;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`
+
+export const StyledButtonVideos = styled.section`
+  margin-left: 35%;
+`
+
+export const StyledLabelVideoDeleted = styled.label`
+  margin-left: 18%;
+  font-family: Roboto;
+  font-weight: 700;
+  font-size: 16px;
+  color: #FBFCF9;
+`
