--- src/components/ShowVideos/styles.js
@@ -1,21 +1,0 @@
-import styled from 'styled-components'
-
-export const SliderContentItem = styled.section`
-  position: relative;
-  height: 100hv;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`
-
-export const StyledButtonVideos = styled.section`
-  margin-left: 35%;
-`
-
-export const StyledLabelVideoDeleted = styled.label`
-  margin-left: 26%;
-  font-family: Roboto;
-  font-weight: 700;
-  font-size: 16px;
-  color: #FBFCF9;
-`
